if文の追加

+++ app/controllers/items_controller.rb
@@ -40,8 +40,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
+    if user_signed_in? && current_user.id == @item.user_id 
+      if @item.destroy
+        redirect_to root_path
+     end
     end
   end
